refactor(node): extract token subcommands to commands/token.rs
Six cmd_token_* functions + the shared cli_create_token_tx helper move
out of main.rs:
  cmd_token_deploy, _transfer, _burn, _balance, _info, _list
  cli_create_token_tx — assembles TokenOp tx, signs, pushes to mempool

Same pattern as wallet / validator / chain / state / mempool / misc
extractions. Three mutating cmds go through the helper to build +
sign + admit; three read-only cmds hit Blockchain accessors directly.

main.rs drops to 3735 lines. Unused imports TOKEN_OP_ADDRESS / TokenOp
/ Transaction came out alongside; PROTOCOL_TREASURY stays since
cmd_start references it.
